pyproject.toml
[tool.poetry]
name = "waste-management"
version = "0.1.0"
description = "Django web application to manage waste records"
authors = ["Bogdan Fometescu <75015831+BogdanMFometescu@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref="3.7.2"
certifi="2023.7.22"
charset-normalizer="3.1.0"
cloudinary="1.33.0"
Django="5.0.11"
django-cloudinary-storage="0.3.0"
django-cors-headers="4.0.0"
django-environ="0.10.0"
django-filter="23.2"
djangorestframework="3.14.0"
environ="1.0"
gunicorn="22.0.0"
idna="3.7"
Pillow="10.2.0"
pytz="2023.3"
requests="2.31.0"
six="1.16.0"
sqlparse="0.5.0"
tzdata="2023.3"
urllib3="1.26.18"
whitenoise="6.4.0"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
xhtml2pdf = "^0.2.13"


[tool.poetry.group.dev.dependencies]
setuptools = "^78.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
